sync smart-snapshot-1-coordination: apply 8755fc97b5
Source-commit: 8755fc97b5 Review comments for Coordination layer
